client/requirements.txt
boto3
fastapi
uvicorn[standard]
python-jose